Fixed mmm For Maven APK
Maven APK (Why are these in maven anyways?) will crash due to it not
needing to generate a jack file.  This wasn't noticed before as brunch
succeeded, however, doing an mmm on one will cause this to fail. This
adds in logic to not generate a jack file for an apk in maven.

Change-Id: I44036bfd51ae3edd9437b153898b419ffc6ba2d6